+++ CHANGELOG.md
@@ -3,6 +3,19 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [Unreleased] - YYYY-MM-DD
+#### Added
+- Configuration settings specified within `.uiharness.yml` in the root of the project.
+
+#### Changed
+#### Deprecated
+#### Removed
+#### Fixed
+#### Security
+
+
+
+
 ## [3.1.0] - 2016-01-26
 #### Added
 - Relay/GraphQL support.